requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==5.0.1
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.1
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
